wip(data-model): slice the fabric value declarations into data-model/api.ts
Parked. Incomplete on its own: the pattern compiler resolves no bare
specifier, so `commonfabric.d.ts` must first become a generated file that
inlines this module's text. That generation lands separately.
